Connection: Do not try to create Jetpack API endpoints if Jetpack is missing.
If the token is valid, jetpack-connection tries to create Jetpack API endpoints, and fails with a fatal error if Jetpack plugin is missing.
This commit suggests a temporary solution for this problem. If Jetpack plugin cannot be found, we do not try to validate the token, and consider it invalid.
In that case the Connection Manager will not try to create the Jetpack API endpoints, and create the `jetpack.remoteAuthorize` endpoint instead.
This means that API requests to the Jetpack endpoints will not longer fail with a 500 error, and pollute the error log with fatal errors, but fail softly providing the 'Error 400: Endpoint does not exist' response.